NEW: pad in sct_maths instead of msct_image
