+++ scripts/sct_register_multimodal.py
@@ -485,12 +485,11 @@ def register(src, dest, paramreg, param, i_step_str):
         cmd = ('')
 
     elif paramreg.steps[i_step_str].algo.lower() in ants_registration_params:
-        from msct_image import pad_image
         # Pad the destination image (because ants doesn't deform the extremities)
         # N.B. no need to pad if iter = 0
         if not paramreg.steps[i_step_str].iter == '0':
             dest_pad = sct.add_suffix(dest, '_pad')
-            pad_image(dest, dest_pad, padding_z=param.padding)
+            sct.run('sct_maths -i '+dest+' -o '+dest_pad+' -pad 0x0x'+str(param.padding))
             dest = dest_pad
 
         cmd = ('isct_antsRegistration '
