+++ scripts/sct_straighten_spinalcord.py
@@ -576,8 +576,7 @@ class SpinalCordStraightener(object):
             padding_x, padding_y, padding_z = padding, padding, padding
             if nx + padding <= leftright_width:
                 padding_x = leftright_width - padding - nx
-            from msct_image import pad_image
-            pad_image(fname_centerline_orient, 'tmp.centerline_pad.nii.gz', padding_x=padding_x, padding_y=padding_y, padding_z=padding_z)
+            sct.run('sct_maths -i '+fname_centerline_orient+' -o tmp.centerline_pad.nii.gz -pad '+str(padding_x)+'x'+str(padding_y)+'x'+str(padding_z))
 
             # Open padded centerline for reading
             sct.printv('\nOpen padded centerline for reading...', verbose)
